fix(microbot): seasonal handlers + leagues modules
SeasonalTransportHandler registry. invalidateContext. Rs2LeaguesTransport split into LeaguesTransport* types. MicrobotPlugin snapshot import.
